Make protobuf-es work with Jest
There is a known issue where Jest expects us to import files without
mentioning the extension (e.g. "import file" instead of "import file.js").

We use a generator command:
```
bug generate proto
```
to convert our protobuf files defined `/proto` into Typescript types
which will be stored in `/proto-generated`.

However, the generator will be explicit when importing files so it will
generate "import file.js" statements. This triggers an error in Jest.

To get around this we're adding a command which goes through the files
and removes the `.js` suffix.

NB: There is a recommended solution using `moduleNameMapper` in Jest [1]
but I wasn't able to get it to work (and other people have had the same
problem [2] [3]). There isn't any intention to correct this as it's
considered to be according to "ECMAScript module rules".

The good news is that we can use `npm run generate:schemas` to generate the
files and fix the import in one go.

[1]: https://connect.build/docs/web/supported-browsers-and-frameworks/#jest
[2]: https://github.com/bufbuild/protobuf-es/issues/282
[3]: https://github.com/bufbuild/protobuf-es/issues/225