First pass at adding some proto schemas
The files defined here will be used to generate Typescript types so
they will need to be called something different than our existing
types.

For example we're unable to create a `variant-analysis.proto` file
as that would try to create a VariantAnalysis type (which we already
have).

I've chosen to follow our convention for calling protobuf definitions
as "schemas" as I think it fits better than calling them "messages".

This is a first pass at writing schemas for remote queries and variant
analyses.

Local queries are quite complicated as they have a considerably
complicated amount of nested types. As a POC I think this is good enough
for now. We can spend time adding those if we choose protobuf-es.