Run integration tests sequentially
Currently we're running our jest tests in multiple parallel threads.

We can see that CI is experiencing a lot of flakiness where a test
will randomly error with the following message:

```
Unable to write into user settings because the content of the file is newer.
```

These are failing only on CI so we believe at least some of the
flakiness is caused by running our tests in a multi-threaded environment
in CI.

So we experimented with running our jest tests in sequence via the
`--runInBand` option: github/vscode-codeql#1998.

While this meant that our tests would take longer to run (15 mins vs. 4-6mins),
most of them now passed on the first CI run:

- first run (12m48s): https://github.com/github/vscode-codeql/actions/runs/3997178977/jobs/6858164662
- another first run (11m58s): https://github.com/github/vscode-codeql/actions/runs/3997164504/jobs/6858136512

You can see more first-run examples on the commits in this PR: github/vscode-codeql#1998.

For comparison, the first CI run for tests running in parallel sees many more failures:
- first attempt (23m8s) - 10 failed jobs: https://github.com/github/vscode-codeql/actions/runs/3997630502/attempts/1
- second attempt (18m58s) - one failed job: https://github.com/github/vscode-codeql/actions/runs/3997630502/attempts/2
- third attempt - green (6m42s): https://github.com/github/vscode-codeql/actions/runs/3997630502/attempts/2

As you can probably tell, there is a notable reduction in flakiness which means we're able to re-run less
jobs to get to a green state.

This can also reveal other sources of flakiness that we've perhaps been ignoring in the past.

So our proposal is for now to run tests in sequence and think about running them in parallel once we understand more
about the way tests interact when run in parallel.

Co-authored-by: Nora Scheuch <norascheuch@github.com>
Co-authored-by: Anders Starcke <starcke@github.com>