Upgrade ts-node to get rid of CI error
CI is exploding with this error:

```
ERROR in ./src/view/webview.tsx
Module build failed (from ./node_modules/ts-loader/index.js):
[18:38:40] 'compileView' errored after 1.41 s
Error: Debug Failure. False expression: Non-string value passed to `ts.resolveTypeReferenceDirective`, likely by a wrapping package working with an outdated `resolveTypeReferenceDirectives` signature. This is probably not a problem in TS itself.
[18:38:40] Error: Compilation errors detected.
    at resultCb (D:\a\vscode-codeql\vscode-codeql\extensions\ql-vscode\gulpfile.ts\webpack.ts:46:14)
```

It seems to be an issue in older versions of ts-node so I've upgraded to a newer version.