Fix `Unhandled arguments` bug
When we added an extra optional param `isTutorialDatabase` to
`openDatabase` [1] , the tests all passed but there is at least one call
[2] where this triggers an `Unhandled arguments` error.

This is because the method call doesn't know which optional param
we're passing.

Let's revert back to the original method signature where we had just
one optional parameter (the database name) and set it to a special
"CodeQL Tutorial Database" value.

We can then use this to understand that the extension is running in
the codespace and we want to skip creating a skeleton pack for it
since it's the tutorial database.

[1]: https://github.com/github/vscode-codeql/blob/5e51bb57f5f76b57768fb6a3543f153f5752c23c/extensions/ql-vscode/src/local-databases.ts#L610
[2]: https://github.com/github/vscode-codeql/blob/5e51bb57f5f76b57768fb6a3543f153f5752c23c/extensions/ql-vscode/src/databaseFetcher.ts#L257-L262