Don't use unhandle rejection listener in codespace
This was implemented here: [1].

It is throwing up a lot of `Unhandled error: Invalid arguments` errors
without any way to see a stack trace for them.

Since we're not able to debug this from the codespace, I'm choosing to
turn it off in the codespace template as it's preventing us from
running the Code Tour successfully.