Remove `glob-promise` and upgrade to `glob` 9
The `glob` package now uses promises in version 9, so we don't need the
separate `glob-promise` package anymore.

This also updates one call site to use `cwd` instead of `join` to
avoid possible issues due to a breaking change in version 8 which treats
Windows path separators differently. By changing the `cwd`, we should
not run into these issues.

Removing the `resolutions` of `glob-parent` did not change the
`package-lock.json`, so this does not have any effect on the package
version we are using.