Show the user a selection box before downloading database
At the moment, we're always deciding which database to download for the
user for an example query.

We'd like to give them a chance to change the database, so here we're
adding a step where we're showing the user a selection box with the
suggested database pre-filled.

They can choose to type in a different database before continuing the
skeleton generation process.