Change kind input to a dropdown
This changes the kind input from a text field to a dropdown in the data
extensions editor. The supported values for each extensible predicate
are based on what is currently in-scope for the documentation. Other
kinds are not supported.

The supported kinds are now stored on the
`extensiblePredicateDefinitions` to make it easier to add new kinds in
the future.