Download databases into extension storage path
Now that we've figured out how to set the storage path for both Mac &
Windows, we also want to make sure we're consistent with the location
where we save databases.

At the moment, our change will download databases directly in the
workspace folder.

When we call `downloadGitHubDatabase()` in other places outside the
wizard, we provide `ctx.storageUri.fsPath` as the location. [1] [2] [3]

Let's do the same here.

I've tested this on Mac for the codespaces-codeql & starter workspaces.

[1]: https://github.com/github/vscode-codeql/blob/c7bb22c312d4bbe46d90ff0a43e9cff22c6bd5ed/extensions/ql-vscode/src/local-databases-ui.ts#L476
[2]: https://github.com/github/vscode-codeql/blob/c7bb22c312d4bbe46d90ff0a43e9cff22c6bd5ed/extensions/ql-vscode/src/extension.ts#L710
[3]: https://github.com/github/vscode-codeql/blob/c7bb22c312d4bbe46d90ff0a43e9cff22c6bd5ed/extensions/ql-vscode/src/extension.ts#L1120