Filter out databases with errors
When we create a skeleton query, we check whether you already have an
existing database with the same name (e.g. `github/codeql`). If we can't
find one, we also check for an existing database with the same language.

If we find one, we select it instead of downloading a new database.

Here we're filtering out databases with errors.