Fix error when model file does not exist
When you have just created a model file using the quick picker/input
box, the data extension editor will try to read it and fail with an
error message. This adds a check to ensure the model file exists and if
it doesn't, it will not try to read in the file.

This should always be safe since the model file picker will only allow
you to select existing files.