Show model filename in data extensions editor
This adds the model filename to the data extensions editor and will also
allow the user to go to the model file by clicking on the filename.

This also updates the general UI to be somewhat more compact by moving
the modeled percentages to be below the header in 1 line.