Clear pack cache after creating extension pack
After creating a new extension pack, the query server will not pick up
the new pack until the cache is cleared. The CLI server will pick up the
change, so this results in an internal error.

This will clear the pack cache after creating the extension pack to
ensure we don't end up in this inconsistent state and the extension pack
is used for queries.